requirements.txt
discord.py==1.7.3
psycopg2==2.9.3
tensorflow==2.8.0
pandas==1.4.1
SQLAlchemy==1.4.31